fix(server): surface runtime persistence degradation
